+++ Arm.h
@@ -3,9 +3,9 @@
 class Arm : public Robot_Part
 {
 private:
-	int power_consumed;
+	double power_consumed;
 public:
-	Arm(string n, int pn, double w, double c, string d, int pc);
-	int get_power_consumed();
+	Arm(string n, int pn, double w, double c, string d, double pc);
+	double get_power_consumed();
 	string to_string();
 };
